Move handshake to PLatform project
