+++ Runtime.sln
@@ -193,6 +193,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Events.Management", "Source
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Resources", "Source\Resources\Resources.csproj", "{C70D8ED8-A701-4697-BE87-B623E37D8976}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Platform", "Source\Platform\Platform.csproj", "{9F665E30-4E54-40FD-B2BC-1E1782222F7E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1322,6 +1324,18 @@ Global
 		{C70D8ED8-A701-4697-BE87-B623E37D8976}.Release|x64.Build.0 = Release|Any CPU
 		{C70D8ED8-A701-4697-BE87-B623E37D8976}.Release|x86.ActiveCfg = Release|Any CPU
 		{C70D8ED8-A701-4697-BE87-B623E37D8976}.Release|x86.Build.0 = Release|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Debug|x64.Build.0 = Debug|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Debug|x86.Build.0 = Debug|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Release|x64.ActiveCfg = Release|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Release|x64.Build.0 = Release|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Release|x86.ActiveCfg = Release|Any CPU
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{ACACF102-3A61-442F-B202-BDF92F6FD26A} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
@@ -1417,5 +1431,6 @@ Global
 		{D5CD6014-99D7-46EC-A149-9ECCD72687AC} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
 		{493EF936-F918-4A78-8AD9-A6A6F93D761A} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
 		{C70D8ED8-A701-4697-BE87-B623E37D8976} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
+		{9F665E30-4E54-40FD-B2BC-1E1782222F7E} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
 	EndGlobalSection
 EndGlobal
