+++ Source/Platform/Handshake/IVerifyContractsCompatability.cs
@@ -3,7 +3,7 @@
 
 using Dolittle.Runtime.Versioning;
 
-namespace Dolittle.Runtime.Server.Handshake;
+namespace Dolittle.Runtime.Platform.Handshake;
 
 /// <summary>
 /// Defines a system that verifies that the contracts versions of the Head and Runtime are compatible.
