+++ Source/Platform/Platform.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <Import Project="../../default.props" />
+
+    <PropertyGroup>
+        <AssemblyName>Dolittle.Runtime.Platform</AssemblyName>
+        <RootNamespace>Dolittle.Runtime.Platform</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="../Configuration/Configuration.csproj" />
+        <ProjectReference Include="../Services/Services.csproj" />
+    </ItemGroup>
+</Project>
